Add support for subsonic POST form parameters & BasicAuth headers
Opensubsonic introduced official support for POST form parameters, but some
legacy clients and servers did already support this, making it a de-facto
standard.

In order to mitigate HPP vulnerabilities introduced by moving authentication
out of the Subsonic server's hands, this plugin now aims to handle and validate
all common credential sources: for backwards compatibility with legacy Subsonic
clients and servers, it would be impractical to restrict credential sources too
much, but they can at least be validated to ensure that if a client sends
credentials in multiple places, all the credentials are consistent.

This change introduces support for POST form parameters, as well as header
validation for BasicAuth credentials.